--- gpt_index/indices/query/knowledge_graph/__init__.py
@@ -1,5 +1,0 @@
-"""Query classes for KG indices."""
-
-from gpt_index.indices.query.knowledge_graph.query import GPTKGTableQuery
-
-__all__ = ["GPTKGTableQuery"]
